pyproject.toml
[tool.poetry]
name = "moneyball2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.0"
ruff = "^0.11.10"
requests = "^2.32.3"
dotenv = "^0.9.9"
google-cloud-firestore = "^2.20.2"
google-auth = "^2.40.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'